requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_py-xgboost-mutex=2.0=cpu_0
_tflow_select=2.3.0=mkl
absl-py=0.7.1=py37_0
agate=1.6.1=pypi_0
appnope=0.1.0=py37_0
asn1crypto=1.4.0=py_0
astor=0.7.1=py37_0
astroid=2.2.5=py37_0
attrs=19.1.0=py_0
autopep8=1.4.4=py_0
azure-common=1.1.26=pypi_0
azure-core=1.9.0=pypi_0
azure-storage-blob=12.6.0=pypi_0
azure-storage-common=2.1.0=pypi_0
babel=2.9.0=pypi_0
backcall=0.1.0=py37_0
blas=1.0=mkl
bleach=3.1.0=py37_0
bokeh=1.3.4=py37_0
boto=2.49.0=py37_0
boto3=1.10.39=py_0
botocore=1.13.39=py_0
bz2file=0.98=py37_1
c-ares=1.16.1=haf1e3a3_0
ca-certificates=2020.10.14=0
cachetools=4.2.0=pypi_0
certifi=2020.6.20=py37_0
cffi=1.12.3=py37hb5b8e2f_0
chardet=3.0.4=py37hecd8cb5_1003
click=7.1.2=py_0
cloudpickle=1.6.0=py_0
colorama=0.3.9=py37_0
confuse=1.3.0=pypi_0
cryptography=2.7=py37ha12b0ac_0
cycler=0.10.0=py37_0
cytoolz=0.11.0=py37he6e4e01_1
dask-core=2.30.0=py_0
dbt=0.18.1=pypi_0
dbt-bigquery=0.18.1=pypi_0
dbt-core=0.18.1=pypi_0
dbt-postgres=0.18.1=pypi_0
dbt-redshift=0.18.1=pypi_0
dbt-snowflake=0.18.1=pypi_0
dbus=1.13.18=h18a8e69_0
decorator=4.4.0=py37_1
defusedxml=0.5.0=py37_1
dill=0.3.2=py_0
docutils=0.14=py37_0
entrypoints=0.3=py37_0
expat=2.2.10=hb1e8313_2
flask=1.1.2=py_0
freetype=2.10.4=ha233b18_0
future=0.18.2=pypi_0
gast=0.2.2=py37_0
gettext=0.19.8.1=h15daf44_3
glib=2.63.1=hd977a24_0
google-api-core=1.16.0=pypi_0
google-auth=1.24.0=pypi_0
google-cloud-bigquery=1.25.0=pypi_0
google-cloud-core=1.3.0=pypi_0
google-crc32c=1.1.0=pypi_0
google-resumable-media=0.5.1=pypi_0
googleapis-common-protos=1.6.0=pypi_0
grpcio=1.16.1=py37h044775b_1
h5py=2.9.0=py37h3134771_0
hdf5=1.10.4=hfa1e0ec_0
hologram=0.0.10=pypi_0
htmlmin=0.1.12=pypi_0
icu=58.2=h0a44026_3
idna=2.8=py37_0
ijson=2.5.1=pypi_0
imagehash=4.2.0=pypi_0
imageio=2.9.0=py_0
importlib-metadata=3.3.0=pypi_0
intel-openmp=2019.4=233
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.6.1=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=pypi_0
isodate=0.6.0=pypi_0
isort=4.3.21=py37_0
itsdangerous=1.1.0=py37_0
jedi=0.16.0=py37_0
jinja2=2.11.2=py_0
jmespath=0.9.4=py_0
joblib=0.13.2=py37_0
jpeg=9b=he5867d9_2
json-rpc=1.13.0=pypi_0
jsonschema=3.0.1=py37_0
jupyter_client=6.1.7=py_0
jupyter_core=4.6.3=py37_0
jupyterthemes=0.20.0=py_1
keras-applications=1.0.7=py_0
keras-preprocessing=1.0.9=py_1
keyring=21.5.0=pypi_0
kiwisolver=1.1.0=py37h0a44026_0
krb5=1.16.4=hddcf347_0
lazy-object-proxy=1.4.1=py37h1de35cc_0
leather=0.3.3=pypi_0
lesscpy=0.13.0=py_1
libblas=3.8.0=14_mkl
libcblas=3.8.0=14_mkl
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.2.1=h0a44026_1007
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
liblapack=3.8.0=14_mkl
libpng=1.6.37=ha441bb4_0
libpq=11.2=h051b688_0
libprotobuf=3.8.0=hd9629dc_0
libsodium=1.0.18=h1de35cc_0
libtiff=4.1.0=hcb84e12_1
libxgboost=0.90=hb1e8313_1
llvm-openmp=10.0.0=h28b9765_0
llvmlite=0.34.0=pypi_0
logbook=1.5.3=pypi_0
lz4-c=1.9.2=h79c402e_3
markdown=3.1.1=py37_0
markupsafe=1.1.1=py37h1de35cc_0
matplotlib=3.3.3=pypi_0
matplotlib-base=2.2.5=py37h11da6c2_1
mccabe=0.6.1=py37_1
minimal-snowplow-tracker=0.0.2=pypi_0
missingno=0.4.2=pypi_0
mistune=0.8.4=py37h1de35cc_0
mkl=2019.4=233
mkl-service=2.3.0=py37hfbe908c_0
mkl_fft=1.2.0=py37hc64f4ea_0
mkl_random=1.1.1=py37h959d312_0
mock=4.0.2=py_0
msrest=0.6.19=pypi_0
nbconvert=5.5.0=py_0
nbformat=4.4.0=py37_0
ncurses=6.2=h0a44026_1
networkx=2.3=pypi_0
notebook=5.7.8=py37_0
numba=0.51.2=pypi_0
numexpr=2.7.1=pypi_0
numpy=1.16.2=py37hacdab7b_0
numpy-base=1.16.2=py37h6575580_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py37_0
openssl=1.1.1h=haf1e3a3_0
oscrypto=1.2.1=pypi_0
packaging=19.1=py37_0
pandas=1.1.4=pypi_0
pandas-profiling=2.9.0=pypi_0
pandoc=2.11=h0dc7051_0
pandocfilters=1.4.2=py37_1
parsedatetime=2.6=pypi_0
parso=0.5.0=py_0
patsy=0.5.1=py37_0
pcre=8.44=hb1e8313_0
pexpect=4.7.0=py37_0
phik=0.10.0=pypi_0
pickleshare=0.7.5=py37_0
pillow=6.1.0=py37hb68e598_0
pip=20.2.4=py37hecd8cb5_0
ply=3.11=py_1
prometheus_client=0.8.0=py_0
prompt_toolkit=2.0.10=py_0
protobuf=3.11.3=pypi_0
psutil=5.6.3=py37h1de35cc_0
psycopg2=2.7.7=pypi_0
psycopg2-binary=2.8.6=pypi_0
ptyprocess=0.6.0=pyhd3eb1b0_2
py-xgboost=0.90=py37hb1e8313_1
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycodestyle=2.5.0=py37_0
pycparser=2.19=py37_0
pycryptodomex=3.9.9=pypi_0
pygments=2.4.2=py_0
pyjwt=1.7.1=pypi_0
pylint=2.3.1=py37_0
pyopenssl=19.0.0=py37_0
pyparsing=2.4.0=py_0
pyqt=5.9.2=py37h655552a_2
pyrsistent=0.17.3=py37haf1e3a3_0
pysocks=1.7.0=py37_0
python=3.7.7=hfe9666f_0_cpython
python-dateutil=2.8.0=py37_0
python-slugify=4.0.1=pypi_0
python_abi=3.7=1_cp37m
pytimeparse=1.1.8=pypi_0
pytz=2019.1=py_0
pywavelets=1.1.1=py37h10e2902_1
pyyaml=3.13=py37h1de35cc_0
pyzmq=19.0.2=py37hb1e8313_1
qt=5.9.7=h468cd18_1
qtconsole=4.5.1=py_0
readline=7.0=h1de35cc_5
requests=2.23.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rope=0.14.0=py_0
rsa=4.6=pypi_0
s3transfer=0.2.1=py37_0
scikit-image=0.16.2=py37h4f17bb1_0
scikit-learn=0.21.3=py37h27c97d8_0
scikit-plot=0.3.7=py_1
scipy=1.5.4=pypi_0
seaborn=0.11.0=pypi_0
selenium=3.141.0=py37h1de35cc_0
send2trash=1.5.0=py37_0
setuptools=50.3.1=py37hecd8cb5_1
shap=0.35.0=py37h94625e5_0
sip=4.19.8=py37h0a44026_0
six=1.15.0=py_0
snowflake-connector-python=2.2.10=pypi_0
sqlalchemy=1.3.19=py37haf1e3a3_0
sqlalchemy-redshift=0.8.1=pyh9f0ad1d_0
sqlite=3.33.0=hffcf06c_0
sqlparse=0.2.3=pypi_0
statsmodels=0.11.1=py37haf1e3a3_0
tables=3.6.1=pypi_0
tangled-up-in-unicode=0.0.6=pypi_0
tensorboard=1.13.1=py37haf313ee_0
tensorflow=1.13.1=mkl_py37h70c3834_0
tensorflow-base=1.13.1=mkl_py37h66b1bf0_0
tensorflow-estimator=1.13.0=py_0
termcolor=1.1.0=py37_1
terminado=0.8.2=py37_0
testpath=0.4.2=py37_0
text-unidecode=1.3=pypi_0
tk=8.6.10=hb0a8c7a_0
toolz=0.11.1=py_0
tornado=6.0.3=py37h1de35cc_3
tqdm=4.52.0=pypi_0
traitlets=4.3.2=py37_0
typing-extensions=3.7.4.3=pypi_0
urllib3=1.24.3=py37_0
visions=0.5.0=pypi_0
wcwidth=0.1.7=py37_0
webencodings=0.5.1=py37_1
werkzeug=0.16.1=pypi_0
wheel=0.35.1=pyhd3eb1b0_0
widgetsnbextension=3.5.0=py37_0
wrapt=1.11.2=py37h1de35cc_0
xlrd=1.2.0=py37_0
xz=5.2.5=h1de35cc_0
yaml=0.1.7=hc338f04_2
zeromq=4.3.3=hb1e8313_3
zipp=3.4.0=pypi_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
